# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  sku: string;
  name: string;
  slug: string;
  shortDescription: string;
  description: string;
  longDescription: string;
  category: string;
  price: number;
  /** @nullable */
  compareAtPrice: number | null;
  imageUrl: string;
  /** @nullable */
  badge: string | null;
  featured: boolean;
  inventory: number;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  sku?: string;
  /** @minLength 1 */
  name: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  shortDescription: string;
  description?: string;
  /** @minLength 1 */
  longDescription: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 0
     * @nullable
     */
  compareAtPrice?: number | null;
  /** @minLength 1 */
  imageUrl: string;
  /** @nullable */
  badge?: string | null;
  featured?: boolean;
  /** @minimum 0 */
  inventory?: number;
}

export interface ProductUpdate {
  /** @minLength 1 */
  sku?: string;
  /** @minLength 1 */
  name?: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  shortDescription?: string;
  description?: string;
  /** @minLength 1 */
  longDescription?: string;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  price?: number;
  /**
     * @minimum 0
     * @nullable
     */
  compareAtPrice?: number | null;
  /** @minLength 1 */
  imageUrl?: string;
  /** @nullable */
  badge?: string | null;
  featured?: boolean;
  /** @minimum 0 */
  inventory?: number;
}

export interface ProductImportInput {
  /**
     * @minItems 1
     * @maxItems 500
     */
  products: ProductInput[];
}

export interface ProductImportResult {
  imported: number;
  products: Product[];
}

export interface AdminSummary {
  totalProducts: number;
  featuredProducts: number;
  lowStockProducts: number;
  categories: number;
}

export type UploadRequestContentType = typeof UploadRequestContentType[keyof typeof UploadRequestContentType];


export const UploadRequestContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 10485760
     */
  size: number;
  contentType: UploadRequestContentType;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
  imageUrl: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
featured?: boolean;
};