# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Storefront product catalogue
  - name: admin
    description: Catalogue management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List storefront products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
  /products/import:
    post:
      operationId: importProducts
      tags: [admin]
      summary: Import products from a parsed CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductImportInput"
      responses:
        "201":
          description: Imported products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImportResult"
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [admin]
      summary: Request a protected product image upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
  /categories:
    get:
      operationId: listCategories
      tags: [products]
      summary: List product categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get catalogue summary
      responses:
        "200":
          description: Catalogue summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, sku, name, slug, shortDescription, description, longDescription, category, price, compareAtPrice, imageUrl, badge, featured, inventory, createdAt]
      properties:
        id: { type: number }
        sku: { type: string }
        name: { type: string }
        slug: { type: string }
        shortDescription: { type: string }
        description: { type: string }
        longDescription: { type: string }
        category: { type: string }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        imageUrl: { type: string }
        badge: { type: ["string", "null"] }
        featured: { type: boolean }
        inventory: { type: number }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, shortDescription, longDescription, category, price, imageUrl]
      properties:
        sku: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        shortDescription: { type: string, minLength: 1, maxLength: 240 }
        description: { type: string }
        longDescription: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        compareAtPrice: { type: ["number", "null"], minimum: 0 }
        imageUrl: { type: string, minLength: 1 }
        badge: { type: ["string", "null"] }
        featured: { type: boolean, default: false }
        inventory: { type: number, minimum: 0, default: 20 }
    ProductUpdate:
      type: object
      properties:
        sku: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        shortDescription: { type: string, minLength: 1, maxLength: 240 }
        description: { type: string }
        longDescription: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        compareAtPrice: { type: ["number", "null"], minimum: 0 }
        imageUrl: { type: string, minLength: 1 }
        badge: { type: ["string", "null"] }
        featured: { type: boolean }
        inventory: { type: number, minimum: 0 }
    ProductImportInput:
      type: object
      required: [products]
      properties:
        products:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/ProductInput"
    ProductImportResult:
      type: object
      required: [imported, products]
      properties:
        imported: { type: number }
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
    AdminSummary:
      type: object
      required: [totalProducts, featuredProducts, lowStockProducts, categories]
      properties:
        totalProducts: { type: number }
        featuredProducts: { type: number }
        lowStockProducts: { type: number }
        categories: { type: number }
    UploadRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: number, minimum: 1, maximum: 10485760 }
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
    UploadResponse:
      type: object
      required: [uploadURL, objectPath, imageUrl]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
        imageUrl: { type: string }