# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  sku: string;
  name: string;
  slug: string;
  shortDescription: string;
  description: string;
  longDescription: string;
  category: string;
  price: number;
  /** @nullable */
  compareAtPrice: number | null;
  imageUrl: string;
  /** @nullable */
  badge: string | null;
  featured: boolean;
  inventory: number;
  createdAt: Date;
}