# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  sku?: string;
  /** @minLength 1 */
  name: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  shortDescription: string;
  description?: string;
  /** @minLength 1 */
  longDescription: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 0
     * @nullable
     */
  compareAtPrice?: number | null;
  /** @minLength 1 */
  imageUrl: string;
  /** @nullable */
  badge?: string | null;
  featured?: boolean;
  /** @minimum 0 */
  inventory?: number;
}